% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.validation.R
\name{cross.validation}
\alias{cross.validation}
\title{cross validation}
\usage{
cross.validation(pre = "", nfold, nthread, nrounds, testparname = NA,
  testpar = NA, plot = T, pdim = 10, width = 1.5)
}
\arguments{
\item{nfold}{how many folds of dataset}

\item{nthread}{number of CPU cores to use, NA for all}

\item{nrounds}{number of training rounds}

\item{plot}{should results be plotted, default=TRUE}

\item{pdim}{dimension of plot, default=10}

\item{width}{proportion of width to height, default=1.5}

\item{defpar}{default parameter:default = list(eta=0.3, gamma=0, max_depth=6, min_child_weight=1, subsample=1, colsample_bytree=1)}
}
\description{
calculates cross validation for given parameter
}
