% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.parameter.test.R
\name{cv.parameter.test}
\alias{cv.parameter.test}
\title{cross validation parameter test}
\usage{
cv.parameter.test(testpar, nfold, nthread, nrounds = 200, plot = T,
  pdim = 10, width = 1.5)
}
\arguments{
\item{testpar}{e.g. list(eta = 0.1, 0.2, ...)}

\item{nfold}{how many folds of dataset}

\item{nthread}{number of CPU cores to use, NA for all}

\item{nrounds}{number of training rounds}

\item{plot}{should results be plotted, default=TRUE}

\item{pdim}{dimension of plot, default=10}

\item{width}{proportion of width to height, default=1.5}
}
\description{
calculates cross validation for given parameter and a testparameter
}
