% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selection.plotter.R
\name{feature.selection.plotter}
\alias{feature.selection.plotter}
\title{plots accuracy loss from feature.selection}
\usage{
feature.selection.plotter(decreasing = F, pdim = 10, width = 1.5,
  saveplot = T, xlim = NA, ylim = NA)
}
\arguments{
\item{pdim}{dimension of plot, default=10}

\item{width}{proportion of width to height, default=1.5}

\item{saveplot}{save plot?(T/F) default = T}

\item{xlim}{zoom on x-axis, e.g. xlim = 280:300}

\item{ylim}{zoom an y-axis, e.g. ylim = c(0,0.25). Use c() with other than natural numbers}
}
\description{
plots accuracy while features are dismissed iteratively
}
