% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selection.R
\name{feature.selection}
\alias{feature.selection}
\title{Feature Selection.}
\usage{
feature.selection(output_dir = "./featureSelection", data = "./data",
  decreasing = F, ndismiss = NA, nrounds = 20, nthread = 0, eta = NA,
  max_depth = NA, savemode = T)
}
\arguments{
\item{output_dir}{output directory}

\item{decreasing}{decreasing = F :least important feature will be dismissed}

\item{ndismiss}{how many features (coordinates) should be dismissed? Use just when \code{decreasing=TRUE}.}

\item{nrounds}{XGBoost parameter: training rounds}

\item{nthread}{0 for all}

\item{eta}{XGBoost parameter: learning rate}

\item{max_depth}{XGBoost parameter: maximum tree depth}

\item{savemode}{save xgb models (T/F)}

\item{impdata}{file to import data parameter}
}
\description{
Dismisses features iteratively by their importance and monitor classification
accuracy. \cr
Trains iteratively a xgb model while dismissing features by importance and
monitoring accuracy. Decreasing will dismiss the most important feature each
round, increasing (default) the least important. The file 'feature.importance'
will contain an ordered list of all dismissed features. \cr
The parameter values used for feature selection are written to the file
'parameter'.
Additionally, in savemode = T, the following files will be created each round i:
\itemize{
  \item round<i>.model, where the trained model is saved.
  \item round<i>.prediction, where the predictions are saved.
  \item round<i>.importance, where the feature importanec is saved.
}
}
