% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{get.default.parameter}
\alias{get.default.parameter}
\title{Default parameters.}
\usage{
get.default.parameter()
}
\value{
named list of default parameters
}
\description{
Default parameters used for training the model.
}
