% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.num.class.R
\name{get.num.class}
\alias{get.num.class}
\title{get number of classes}
\usage{
get.num.class(M1, M2 = NA)
}
\arguments{
\item{M1}{first xgb.Dmatrix}

\item{M2}{second xgb.Dmatrix}
}
\description{
gets maximum class from one or two xgb.Dmatrix elements and adds 1
}
