% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{get.parameter}
\alias{get.parameter}
\title{Get parameter.}
\usage{
get.parameter(params, parameter)
}
\arguments{
\item{parameter}{parameters}
}
\description{
Read 'parameter' file created with \link{set.parameter} and return requested
parameters.
}
