% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.data.R
\name{import.data}
\alias{import.data}
\title{import data}
\usage{
import.data(savefolder, crd, sts, label, trainsplit)
}
\arguments{
\item{savefolder}{output folder}

\item{crd}{coordinate file to import, e.g. dihedrals}

\item{sts}{states file to import that match coordinates}

\item{label}{coordinate label, as file link or 'dihedrals' for Phi2, Psi2, Phi3, ...}

\item{trainsplit}{how much of the data for training, rest for test}
}
\description{
import data
}
\details{
This function reads the files \code(crd), \code(sts) and \code(label) as coordinates, states and labels.
The labels represent the column names of the coordinates and the states should match the rows of the coordinates.
The data then is separated in a training and a test set and saved in the xgboost matrix format.
The following files are created:
\itemize{
  \item all.xgb.Dmatrix, where the whole data set is saved.
  \item train.xgb.Dmatrix, where the training set is saved.
  \item test.xgb.Dmatrix, where the test set is saved.
  \item train.index, where the train indix are saved.
  \item test.index, where the test index is saved.
  \item data.import.parameter, where the input parameter are saved to use by other functions.
}
}
\examples{
import.data("./file.dih", "./macrostates", "./file.dih.names", "./savefolder", 0.75)
import.data("./file.dih", "./macrostates", "dihedrals", "./savefolder", 0.5)
}
