% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.importance.R
\name{order.importance}
\alias{order.importance}
\title{order importance}
\usage{
order.importance(M, decreasing = F)
}
\arguments{
\item{M}{importance matrix of model}

\item{decreasing}{if FALSE (default) least important feature is on top}
}
\description{
order importance in dependence of population
}
