% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.test.R
\name{parameter.test}
\alias{parameter.test}
\title{parameter test}
\usage{
parameter.test(nthread, nrounds, defpar = list(objective = "multi:softmax",
  num_class = 2, eta = 0.3, gamma = 0, max_depth = 6, min_child_weight = 1,
  subsample = 1, colsample_bytree = 1), testpar, plot = T, pdim = 10,
  width = 1.5)
}
\arguments{
\item{nthread}{number of CPU cores to use, NA for all}

\item{nrounds}{number of training rounds}

\item{defpar}{default parameter, default = list(eta=0.3, gamma=0, max_depth=6, min_child_weight=1, subsample=1, colsample_bytree=1)}

\item{testpar}{list of parameter to test: data.frame(parameter name = c(start, stop, stepsize, adjust nrounds(T/F)), eta = c(0.1,0.5,0.1,T))}

\item{plot}{should results be plotted, default=TRUE}

\item{pdim}{dimension of plot, default=10}

\item{width}{proportion of width to height, default=1.5}
}
\description{
parameter test
}
