% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.single.class.importance.R
\name{plt.sci}
\alias{plt.sci}
\title{makes sci plot}
\usage{
plt.sci(M, savename, pdim, width)
}
\arguments{
\item{M}{importance matrix of model}

\item{pdim}{plot dimension (height)}

\item{width}{plot width (pdim*width)}
}
\description{
makes sci plot
}
