% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.class.importance.R
\name{single.class.importance}
\alias{single.class.importance}
\title{single class importance}
\usage{
single.class.importance(model = NA, x)
}
\arguments{
\item{x}{class for which importance should be calculated}
}
\description{
calculates importance of single class from xgb.model
}
