% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.model.R
\name{train.model}
\alias{train.model}
\title{Train xgboost model.}
\usage{
train.model(data_dir, output_dir, params = NA, nrounds = 20)
}
\arguments{
\item{data_dir}{directory containing the files train.parameter,
feature.names, train.xgb.Dmatrix, test.xgb.Dmatrix (see \link{\code{import.data}})}

\item{output_dir}{directory to which results are written}

\item{nrounds}{number of training rounds}
}
\description{
TODO: detailed description.
TODO: Parameter, test.matrix (?)
TODO: data_dir
}
