% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.map.R
\name{confusion.map}
\alias{confusion.map}
\title{confusion map}
\usage{
confusion.map(output_dir = "./confusionMap", impdata = "./data", pred,
  average = NULL, noDiagonal = T, pdim = 10, width = 1.5)
}
\arguments{
\item{output_dir}{output directory}

\item{pred}{prediction file}

\item{average}{NA, 'label' or 'prediction'. Default = NA}

\item{noDiagonal}{should diagonal not be plotted for better visibility of missclassification? default = T}
}
\description{
makes a confusion matrix and plot for missclassification
}
