% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.map.R
\name{confusion.map}
\alias{confusion.map}
\title{confusion map}
\usage{
confusion.map(dir = "/confusionMap", pred, average = NULL, noDiagonal = T)
}
\arguments{
\item{dir}{output directory}

\item{pred}{prediction file}

\item{average}{NA, 'label' or 'prediction'. Default = NA}

\item{noDiagonal}{should diagonal not be plotted for better visibility of missclassification? default = T}
}
\description{
makes a confusion matrix and plot for missclassification
}
