% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.data.R
\name{import.data}
\alias{import.data}
\title{Prepare training and test data.}
\usage{
import.data(output_dir, coords, states, labels, trainsplit)
}
\arguments{
\item{output_dir}{Character, name of the output directory.}

\item{coords}{Character, name of the coordinates file}

\item{states}{Character, name of the states file}

\item{trainsplit}{Numeric, fraction of the data that should be used for
training, in [0,1]}

\item{label}{Character, filename or 'dihedrals' to get Phi2, Psi2, Phi3, ...}
}
\value{
named list with elements: num_class
}
\description{
Reads the files \code{coords}, \code{states} and \code{labels} as coordinates,
states and labels. The labels represent the column names of the coordinates
and the states should match the rows of the coordinates. The data then is
separated in a training and a test set and saved in the xgboost matrix format.
The following files are created:
\itemize{
  \item all.xgb.Dmatrix, where the whole data set is saved.
  \item train.xgb.Dmatrix, where the training set is saved.
  \item test.xgb.Dmatrix, where the test set is saved.
  \item train.index, where the train indix are saved.
  \item test.index, where the test index is saved.
  \item data.import.parameter, where the input parameter are saved to be used
        by other functions.
}
}
\examples{
import.data("./savefolder","./file.dih", "./macrostates", "./file.dih.names", 0.75)
import.data("./savefolder", "./file.dih", "./macrostates", "dihedrals", 0.5)
}
