% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.sci.R
\name{plt.sci}
\alias{plt.sci}
\title{single class importance plotter}
\usage{
plt.sci(M, savename, pdim, width)
}
\arguments{
\item{M}{importance matrix of model}

\item{pdim}{plot dimension (height)}

\item{width}{plot width (pdim*width)}
}
\description{
plots importance from single.class.importance
}
