% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.single.class.importance.R
\name{plt.single.class.importance}
\alias{plt.single.class.importance}
\title{Single class importance plot}
\usage{
plt.single.class.importance(pre = "singleClassImportance/sci", model = NA,
  names = NA, colmin = NA, nfeatures = NA, pdim = 10, width = 1)
}
\arguments{
\item{model}{directory to xgb.model file. If NA (default) model from ./model/xgb.model is taken.}

\item{colmin}{Sets minimum of imnportance a feature has to contribute to at least one class. Can be set as vector to produce multiple plots: colmin = c(0.1,0.2,0.3)}

\item{nfeatures}{Returns plot with 'nfeatures' most important features. Can be set as vector to produce multiple plots: c(4,5,6)}

\item{pdim}{plot dimension (height = pdim)}

\item{width}{plot width (pdim*width)}
}
\description{
Plot the feature importance per state. Feature importance is normalized. \cr
Specify \code{colmin} to obtain a plot containing all coordinates with
a feature importance larger than this minimum importance. \cr
Specify \code{nfeatures} to obtain a plot containing the n most important
coordinates.
TODO update parameter descriptions
}
