% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.single.class.importance.R
\name{plt.single.class.importance}
\alias{plt.single.class.importance}
\title{single class importance plotter}
\usage{
plt.single.class.importance(model = NA, colmin = 0, pdim = 10,
  width = 1)
}
\description{
plots importance from single.class.importance
}
