% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{set.parameter}
\alias{set.parameter}
\title{Set parameter.}
\usage{
set.parameter(parameter = list())
}
\arguments{
\item{parameter}{list of parameters (for all parameters  not included in this
list, the default values are used)}
}
\description{
Write (custom) parameters to file 'parameters'. To use the default parameters
call this function without any argument.
}
