% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.parameter.R
\name{set.parameter}
\alias{set.parameter}
\title{set parameter}
\usage{
set.parameter(parameter)
}
\arguments{
\item{parameter}{list of parameters, e.g. list(eta = 0.3, max_depth = 6, ...)}
}
\description{
set parameter
}
