% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.model.R
\name{train.model}
\alias{train.model}
\title{Train xgboost model.}
\usage{
train.model(data_dir, output_dir, params = NA, nrounds = 20)
}
\arguments{
\item{nrounds}{TODO why is this not taken from parameter file}

\item{num_class}{TODO why is this not taken from parameter file}
}
\description{
TODO: detailed description.
TODO: Parameter, test.matrix (?)
}
